requirements.txt
fastapi==0.111.0
uvicorn==0.19.0
python-dotenv==1.0.1
requests==2.31.0
pydantic==2.9.2
crewai